feat/send timeout (#42)
* convert tabs to spaces

* add send_timeout parameter

For backwards compatibility, if the user gives a `read_timeout` but no
`send_timeout`, we assume they want a `send_timeout` matching their
`read_timeout`. If neither are given, the defaults apply.

* fallback send_timeout in all cases

We must do this for connector level config, as well as connection level
config. Also moved dsn fallbacks into a utility function.

* format readme for better source readability

* document send_timeout

* add tests for send timeout backwards compatibility

* suppress socket errors

There are spurious socket errors in the logs when using `tcpsock:settimeouts` which only appear in certain environments and are not related to the socket being tested. The socket is not in an error state, and all of our tests pass with proper checks for errors. So we now suppress the built in errors, as recommended in the lua-nginx-module docs.

* bump nginx to 1.19.9

Something is broken in the openresty patch set, but modules using 1.19.9
seem to build in travis fine.